main pilot WIP
