+++ app/Livewire/SubmissionsTableView.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Livewire;
+
+use Livewire\Component;
+
+class SubmissionsTableView extends Component
+{
+    public function render()
+    {
+        return view('livewire.submissions-table-view');
+    }
+}
