+++ resources/views/livewire/submissions-table-view.blade.php
@@ -1,0 +1,3 @@
+<div>
+    {{-- Knowing others is intelligence; knowing yourself is true wisdom. --}}
+</div>
