+++ resources/views/livewire/xlsforms-table-view.blade.php
@@ -1,0 +1,3 @@
+<div>
+    {{ $this->table }}
+</div>
